Added Linear Regression
Finally figured a way to add a linear regression table for any number of polynomial values. This code outputs the regression tables to a text file that is already setup for latex use as the output is in latex. Both the polynomial number and latex file name are able to be entered by the user. HOWEVER, there are no checks to make sure the file or value the user enters are correct so if the wrong input is entered the program will just crash.